fix(select instance): recreate connector when selecting an instance
The API connector is tied to a specific hub, so when selecting an
instance we need to make sure we re-create the connector to point to the
right hub. Otherwise our requests end up going to the wrong hub and we
get a 403 back with a confusing response "Permission denied when getting
Evo license access". TODO: update that in ps-auth-lib.

This code is bad because it accesses a private attribute. But there
doesn't look to be an obvious way to change the base URL. Also the whole
function should be moved into evo_context instead of being on workspace
tools.